refactor(tab-group): extract terminal opt-out empty state
